Switch to action hook add_attachment
I still haven't been able to reproduce the issue at hand, in which
duplicate posts were being created automatically on an end user's site,
but after exploring the current code, I don't feel right about using the
filter for wp_update_attachment_metadata, as it seems like it could
actually fire more often than we wish. It's also meant to be more of a
filter, and less of an instance where we should be performing DB actions,
so I feel much better using add_attachment instead.

The only concern with add_attachment was wether meta data about the
uploaded image would make it through, but after exploring the core code,
it appears that all meta is taken care of before the add_attachment action
is called. I've tested thoroughly with 3.3.1, so I'm going to call it
solved and at the least freshened up for a minor release unless the end
user can get back to me with more specific data and a reproducable
scenario.