Experimenting with adding the image to content
This is turning out tougher than it's worth though as there's no perfect
hook for 'thumbnails have been created', so any time we add an image to
the content, it's the full size image. At this point I think I'm going to
move this stuff to a new branch for development, then tie up the new post
format feature and call it a day.