Fix bugs related to post formats
Somehow I let a bunch of bugs slip through when post formats were not declared as supported in a theme, or when post formats were supported, but none were yet registered.

Doesn't appear that anyone has complained of the issue, which may say something about how often settings are changed, but this will take care of it either way.