Bail when an image is uploaded through an existing post
If an image is uploaded through an existing post, it seems weird to create another post out of it and then assign it as a featured image.

Now, we detect this scenario and bail out if a post parent is detected. The decision is passed onto the user through a filter - afip_post_parent_continue